requirements.txt
asgiref==3.2.7
Django==3.0.6
django-autocomplete-light==3.5.0
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
django-js-asset==1.2.2
django-mptt==0.11.0
